Rename package routing.filter.container to routing.container
Completes the Filter->Transformer rename: the package held only container
transformers, so the now-misleading 'filter' level is dropped. Updates the
JPMS module export accordingly.
