+++ jpro-routing/example/src/main/java/example/filters/SimpleHamburgerMenu.java
@@ -10,7 +10,7 @@ import javafx.scene.control.Separator;
 import javafx.scene.layout.*;
 import one.jpro.platform.routing.LinkUtil;
 import one.jpro.platform.routing.Request;
-import one.jpro.platform.routing.filter.container.Container;
+import one.jpro.platform.routing.container.Container;
 
 import java.util.List;
